- Added Web Api Test Endpoints - Added database image and container creation steps - Added Identity Server to (TODO) create SSO Authentication service
Problems to resolve:
- App cannot connect to databse enginge container due to:
Microsoft.Data.SqlClient.SqlException (0x80131904): A network-related or instance-specific error occurred while establishing a connection to SQL Server. The server was not found or was not accessible. Verify that the instance name is correct and that SQL Server is configured to allow remote connections. (provider: TCP Provider, error: 40 - Could not open a connection to SQL Server)